+++ packages/server/core/src/base/middlewares/index.ts
@@ -3,3 +3,4 @@ export * from './monitor';
 export * from './renderHandler';
 export * from './logger';
 export * from './frameworkHeader';
+export { getLoaderCtx } from './customServer';
