+++ packages/server/core/src/base/middlewares/renderHandler/ssrCache.ts
@@ -8,9 +8,10 @@ import type {
   Container,
 } from '@modern-js/types';
 import { createMemoryStorage } from '@modern-js/runtime-utils/storer';
+import { getPathname } from '@modern-js/runtime-utils/request';
+import { createTransformStream } from '../../utils';
 import type { SSRServerContext, ServerRender } from '../../../core/server';
 import { createReadableStreamFromReadable } from '../../adapters/node/polyfills/stream';
-import { createTransformStream, getPathname } from '../../utils';
 
 interface CacheStruct {
   val: string;
