+++ packages/server/core/src/base/middlewares/renderHandler/ssrRender.ts
@@ -6,12 +6,12 @@ import {
 } from '@modern-js/utils/universal/constants';
 import * as isbot from 'isbot';
 import {
-  getRuntimeEnv,
   parseHeaders,
   parseQuery,
   getHost,
   getPathname,
-} from '../../utils';
+} from '@modern-js/runtime-utils/request';
+import { getRuntimeEnv } from '../../utils';
 import {
   SSRServerContext,
   ServerManifest,
