+++ packages/server/core/src/base/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './env';
-export * from './request';
 export * from './transformStream';
 export * from './middlewareCollector';
 export * from './error';
